refactor: remove folder and a380 header
