+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { A320Header } from '../components/home/headers/A320Header';
+import { A320Header } from '../components/home/A320Header';
 import { Installer } from '../components/home/Installer';
 import { Community } from '../components/home/Community';
 import { PartnerSection } from '../components/home/PartnerSection';
